LIBITD-434. Moved duplicated "sort_and_paginate_results" method into helper module.
https://issues.umd.edu/browse/LIBITD-434
